backend/requirements.txt
selenium>=4.0.0
beautifulsoup4>=4.9.0
requests>=2.25.0
flask>=2.0.0
python-dotenv>=0.19.0
lxml>=4.6.0
flask-cors>=3.0.10
openai>=1.0.0
